+++ src/net/Qt.NetCore.sln
@@ -1,34 +1,57 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
-VisualStudioVersion = 15.0.26403.7
+VisualStudioVersion = 15.0.27428.2011
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Qt.NetCore", "Qt.NetCore\Qt.NetCore.csproj", "{7CB05369-9F78-4FC7-AE17-3614998553D1}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Qt.NetCore", "Qt.NetCore\Qt.NetCore.csproj", "{D87FE1E6-FCDB-47C1-8F82-4C5871FBB30D}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Qt.NetCore.Sandbox", "Qt.NetCore.Sandbox\Qt.NetCore.Sandbox.csproj", "{2AB6D1AB-FCC0-4E8B-B219-82915B4476B8}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Qt.NetCore.Tests", "Qt.NetCore.Tests\Qt.NetCore.Tests.csproj", "{B0C83887-D424-4747-BF0A-E8AA47BE1548}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Qt.NetCore.Sandbox", "Qt.NetCore.Sandbox\Qt.NetCore.Sandbox.csproj", "{4218FE2F-4956-4D1A-9711-7FD8C1CFA724}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
+		Linux-arm-Debug|Any CPU = Linux-arm-Debug|Any CPU
+		Linux-arm-Release|Any CPU = Linux-arm-Release|Any CPU
+		Linux-x64-Debug|Any CPU = Linux-x64-Debug|Any CPU
+		Linux-x64-Release|Any CPU = Linux-x64-Release|Any CPU
+		Win-x64-Debug|Any CPU = Win-x64-Debug|Any CPU
+		Win-x64-Release|Any CPU = Win-x64-Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{7CB05369-9F78-4FC7-AE17-3614998553D1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7CB05369-9F78-4FC7-AE17-3614998553D1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7CB05369-9F78-4FC7-AE17-3614998553D1}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7CB05369-9F78-4FC7-AE17-3614998553D1}.Release|Any CPU.Build.0 = Release|Any CPU
-		{2AB6D1AB-FCC0-4E8B-B219-82915B4476B8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2AB6D1AB-FCC0-4E8B-B219-82915B4476B8}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2AB6D1AB-FCC0-4E8B-B219-82915B4476B8}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2AB6D1AB-FCC0-4E8B-B219-82915B4476B8}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B0C83887-D424-4747-BF0A-E8AA47BE1548}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B0C83887-D424-4747-BF0A-E8AA47BE1548}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B0C83887-D424-4747-BF0A-E8AA47BE1548}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B0C83887-D424-4747-BF0A-E8AA47BE1548}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D87FE1E6-FCDB-47C1-8F82-4C5871FBB30D}.Linux-arm-Debug|Any CPU.ActiveCfg = Linux-arm-Debug|Any CPU
+		{D87FE1E6-FCDB-47C1-8F82-4C5871FBB30D}.Linux-arm-Debug|Any CPU.Build.0 = Linux-arm-Debug|Any CPU
+		{D87FE1E6-FCDB-47C1-8F82-4C5871FBB30D}.Linux-arm-Release|Any CPU.ActiveCfg = Linux-arm-Release|Any CPU
+		{D87FE1E6-FCDB-47C1-8F82-4C5871FBB30D}.Linux-arm-Release|Any CPU.Build.0 = Linux-arm-Release|Any CPU
+		{D87FE1E6-FCDB-47C1-8F82-4C5871FBB30D}.Linux-x64-Debug|Any CPU.ActiveCfg = Linux-x64-Debug|Any CPU
+		{D87FE1E6-FCDB-47C1-8F82-4C5871FBB30D}.Linux-x64-Debug|Any CPU.Build.0 = Linux-x64-Debug|Any CPU
+		{D87FE1E6-FCDB-47C1-8F82-4C5871FBB30D}.Linux-x64-Release|Any CPU.ActiveCfg = Linux-x64-Release|Any CPU
+		{D87FE1E6-FCDB-47C1-8F82-4C5871FBB30D}.Linux-x64-Release|Any CPU.Build.0 = Linux-x64-Release|Any CPU
+		{D87FE1E6-FCDB-47C1-8F82-4C5871FBB30D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D87FE1E6-FCDB-47C1-8F82-4C5871FBB30D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D87FE1E6-FCDB-47C1-8F82-4C5871FBB30D}.Win-x64-Debug|Any CPU.ActiveCfg = Win-x64-Debug|Any CPU
+		{D87FE1E6-FCDB-47C1-8F82-4C5871FBB30D}.Win-x64-Debug|Any CPU.Build.0 = Win-x64-Debug|Any CPU
+		{D87FE1E6-FCDB-47C1-8F82-4C5871FBB30D}.Win-x64-Release|Any CPU.ActiveCfg = Win-x64-Release|Any CPU
+		{D87FE1E6-FCDB-47C1-8F82-4C5871FBB30D}.Win-x64-Release|Any CPU.Build.0 = Win-x64-Release|Any CPU
+		{4218FE2F-4956-4D1A-9711-7FD8C1CFA724}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4218FE2F-4956-4D1A-9711-7FD8C1CFA724}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4218FE2F-4956-4D1A-9711-7FD8C1CFA724}.Linux-arm-Debug|Any CPU.ActiveCfg = Linux-arm-Debug|Any CPU
+		{4218FE2F-4956-4D1A-9711-7FD8C1CFA724}.Linux-arm-Debug|Any CPU.Build.0 = Linux-arm-Debug|Any CPU
+		{4218FE2F-4956-4D1A-9711-7FD8C1CFA724}.Linux-arm-Release|Any CPU.ActiveCfg = Linux-arm-Release|Any CPU
+		{4218FE2F-4956-4D1A-9711-7FD8C1CFA724}.Linux-arm-Release|Any CPU.Build.0 = Linux-arm-Release|Any CPU
+		{4218FE2F-4956-4D1A-9711-7FD8C1CFA724}.Linux-x64-Debug|Any CPU.ActiveCfg = Linux-x64-Debug|Any CPU
+		{4218FE2F-4956-4D1A-9711-7FD8C1CFA724}.Linux-x64-Debug|Any CPU.Build.0 = Linux-x64-Debug|Any CPU
+		{4218FE2F-4956-4D1A-9711-7FD8C1CFA724}.Linux-x64-Release|Any CPU.ActiveCfg = Linux-x64-Release|Any CPU
+		{4218FE2F-4956-4D1A-9711-7FD8C1CFA724}.Linux-x64-Release|Any CPU.Build.0 = Linux-x64-Release|Any CPU
+		{4218FE2F-4956-4D1A-9711-7FD8C1CFA724}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4218FE2F-4956-4D1A-9711-7FD8C1CFA724}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4218FE2F-4956-4D1A-9711-7FD8C1CFA724}.Win-x64-Debug|Any CPU.ActiveCfg = Win-x64-Debug|Any CPU
+		{4218FE2F-4956-4D1A-9711-7FD8C1CFA724}.Win-x64-Debug|Any CPU.Build.0 = Win-x64-Debug|Any CPU
+		{4218FE2F-4956-4D1A-9711-7FD8C1CFA724}.Win-x64-Release|Any CPU.ActiveCfg = Win-x64-Release|Any CPU
+		{4218FE2F-4956-4D1A-9711-7FD8C1CFA724}.Win-x64-Release|Any CPU.Build.0 = Win-x64-Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {C93464A0-DFE4-4C7F-8975-EBDECC410ACA}
 	EndGlobalSection
 EndGlobal
