+++ src/net/Qt.NetCore/IUiContext.cs
