+++ src/net/Qt.NetCore/QGuiApplication.cs
