+++ src/net/Qt.NetCore/Qt.NetCore.csproj
@@ -1,59 +1,63 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="15.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
+<Project Sdk="Microsoft.NET.Sdk">
+
   <PropertyGroup>
-    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
-    <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{7CB05369-9F78-4FC7-AE17-3614998553D1}</ProjectGuid>
-    <OutputType>Library</OutputType>
-    <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>Qt.NetCore</RootNamespace>
-    <AssemblyName>Qt.NetCore</AssemblyName>
-    <TargetFrameworkVersion>v4.5.2</TargetFrameworkVersion>
-    <FileAlignment>512</FileAlignment>
+    <TargetFramework>netstandard2.0</TargetFramework>
+    <Configurations>Win-x64-Debug;Win-x64-Release;Linux-x64-Debug;Linux-x64-Release;Linux-arm-Debug;Linux-arm-Release</Configurations>
   </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
-    <DebugSymbols>true</DebugSymbols>
-    <DebugType>full</DebugType>
-    <Optimize>false</Optimize>
-    <OutputPath>bin\Debug\</OutputPath>
-    <DefineConstants>DEBUG;TRACE</DefineConstants>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
-    <DebugType>pdbonly</DebugType>
-    <Optimize>true</Optimize>
-    <OutputPath>bin\Release\</OutputPath>
-    <DefineConstants>TRACE</DefineConstants>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-  </PropertyGroup>
+
   <ItemGroup>
-    <Reference Include="System" />
-    <Reference Include="System.Core" />
-    <Reference Include="System.Xml.Linq" />
-    <Reference Include="System.Data.DataSetExtensions" />
-    <Reference Include="Microsoft.CSharp" />
-    <Reference Include="System.Data" />
-    <Reference Include="System.Net.Http" />
-    <Reference Include="System.Xml" />
+    <Compile Include="..\..\interop\Interop.cs" Link="Interop.cs" />
   </ItemGroup>
+
   <ItemGroup>
-    <Compile Include="..\..\interop\Interop.cs">
-      <Link>Interop.cs</Link>
-    </Compile>
-    <Compile Include="Callback.cs" />
-    <Compile Include="Helpers.cs" />
-    <Compile Include="ITypeCreator.cs" />
-    <Compile Include="NetTypeInfoManager.cs" />
-    <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="QQmlApplicationEngine.cs" />
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="debug-variables.txt" Condition=" '$(Configuration)' == 'Debug' ">
+    <None Update="debug-variables.txt">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
   </ItemGroup>
-  <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
+
+  <Choose>
+    <When Condition=" '$(Configuration)'=='Win-x64-Debug'">
+      <ItemGroup>
+        <None Include="..\..\native\win-x64\debug\QtNetCoreQml.dll" Link="QtNetCoreQml.dll">
+          <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+        </None>
+      </ItemGroup>
+    </When>
+    <When Condition=" '$(Configuration)'=='Win-x64-Release'">
+      <ItemGroup>
+        <None Include="..\..\native\win-x64\release\QtNetCoreQml.dll" Link="QtNetCoreQml.dll">
+          <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+        </None>
+      </ItemGroup>
+    </When>
+    <When Condition=" '$(Configuration)'=='Linux-x64-Debug'">
+      <ItemGroup>
+        <None Include="..\..\native\linux-x64\debug\libQtNetCoreQml.so.1.0.0" Link="libQtNetCoreQml.so">
+          <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+        </None>
+      </ItemGroup>
+    </When>
+    <When Condition=" '$(Configuration)'=='Linux-x64-Release'">
+      <ItemGroup>
+        <None Include="..\..\native\linux-x64\release\libQtNetCoreQml.so.1.0.0" Link="libQtNetCoreQml.so">
+          <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+        </None>
+      </ItemGroup>
+    </When>
+    <When Condition=" '$(Configuration)'=='Linux-arm-Debug'">
+      <ItemGroup>
+        <None Include="..\..\native\linux-arm\debug\libQtNetCoreQml.so.1.0.0" Link="libQtNetCoreQml.so">
+          <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+        </None>
+      </ItemGroup>
+    </When>
+    <When Condition=" '$(Configuration)'=='Linux-arm-Release'">
+      <ItemGroup>
+        <None Include="..\..\native\linux-arm\release\libQtNetCoreQml.so.1.0.0" Link="libQtNetCoreQml.so">
+          <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+        </None>
+      </ItemGroup>
+    </When>
+  </Choose>
+
 </Project>
