+++ src/net/Qt.NetCore/Utils.cs
