--- src/netCore/Qt.NetCore.Sandbox/Page1.qml
@@ -1,7 +1,0 @@
-import QtQuick 2.7
-
-Page1Form {
-//    button1.onClicked: {
-//        console.log("Button Pressed. Entered text: " + textField1.text);
-//    }
-}
