--- src/netCore/Qt.NetCore/ITypeCreator.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Qt.NetCore
-{
-    public interface ITypeCreator
-    {
-        object Create(Type type);
-    }
-}
